refactor: extract desktop agent prompts and helpers
