+++ routes/agent-run-memory-persist.ts
@@ -4,7 +4,7 @@ import {
   extractProjectKnowledge,
   compactConversationMemory,
 } from '../agent/core/memory.ts';
-import { summarizeText } from '../agent/core/ai-client.ts';
+import { summarizeText } from '../agent/core/summarizer.ts';
 import { log } from '../helpers/logger.ts';
 
 export async function persistAgentRunMemory({
